credit-bias/requirements.txt
# local package
-e .

# external requirements
click
colorlog
Sphinx
coverage
awscli
flake8
pandas
scikit-learn
python-dotenv>=0.5.1
joblib
xgboost